Move command resolution to module
